# server/services/swagger/api-docs.yaml
openapi: 3.0.0
info:
  title: SGL API
  description: API para el Sistema de Gestión de Laboratorios (SGL). Este sistema está diseñado para facilitar la gestión de equipos, usuarios, reservas, y mantenimiento de laboratorios.
  version: 1.0.0
  contact:
    name: Angelo G
    email: angelogarcia2021@@gmail.com
    url: https://wa.me/593982718318
servers:
- url: http://localhost:3000/api
  description: API local
paths:
...
  /lab/report/pdf:
    get:
      tags:
      - Laboratorios
      summary: Generar reporte PDF de laboratorios
      description: Genera un reporte en formato PDF de los laboratorios dentro de un rango de fechas y zona horaria especificada.
      operationId: generatePdfReport
      parameters:
      - name: startDate
        in: query
        required: false
        description: Fecha de inicio para filtrar los laboratorios
        schema:
          type: string
          format: date
          example: "2024-01-01"
      - name: endDate
        in: query
        required: false
        description: Fecha de finalización para filtrar los laboratorios
        schema:
          type: string
          format: date
          example: "2024-12-31"
      responses:
        '200':
          description: Reporte PDF generado exitosamente
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: "Parámetros inválidos."
        '404':
          description: No se encontraron registros de laboratorios para generar el reporte
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: "No se encontraron registros de laboratorios para generar el reporte."
        '500':
          description: Error interno en el servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: "Error al generar el reporte PDF de laboratorios."
  # ACCESS LAB 
  /access-lab/all:
    get:
      tags:
      - Accesos a Laboratorios
      summary: Obtener lista de accesos a laboratorios
      description: Obtiene una lista de los accesos a laboratorios con paginación, filtrado por número de tarjeta de identificación.
      operationId: getAllAccessLabs
      parameters:
      - name: page
        in: query
        required: false
        description: "Número de página para paginación (valor por defecto: 1)"
        schema:
          type: integer
          example: 1
      - name: limit
        in: query
        required: false
        description: " Número de registros por página (valor por defecto: 10)"
        schema:
          type: integer
          example: 10
      - name: identification_card
        in: query
        required: false
        description: Filtro por número de tarjeta de identificación
        schema:
          type: string
      responses:
        '200':
          description: Accesos a laboratorios obtenidos exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "Accesos a laboratorios obtenidos exitosamente."
                  data:
                    type: object
                    properties:
                      totalRecords:
                        type: integer
                        example: 50
                      totalPages:
                        type: integer
                        example: 5
                      currentPage:
                        type: integer
                        example: 1
                      recordsPerPage:
                        type: integer
                        example: 10
                      labs:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 1
                            identification_card:
                              type: string
                              example: "12345"
                            status:
                              type: string
                              example: "Active"
                            createdAt:
                              type: string
                              format: date-time
                              example: "2024-12-10T00:00:00Z"
        '404':
          description: No se encontraron accesos a laboratorios
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: "No se encontraron accesos a laboratorios."
        '500':
          description: Error interno en el servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: "Error al obtener accesos a laboratorios."
  /access-lab/get-by-id/{id}:
    get:
      tags:
      - Accesos a Laboratorios
      summary: Obtener acceso a laboratorio por ID
      description: Obtiene los detalles de un acceso a laboratorio específico.
      operationId: getAccessLabById
      parameters:
      - name: id
        in: path
        required: true
        description: ID del acceso al laboratorio a obtener
        schema:
          type: string
          example: 1
      responses:
        '200':
          description: Acceso al laboratorio obtenido exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "Acceso al laboratorio obtenido exitosamente."
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      identification_card:
                        type: string
                        example: "12345"
                      status:
                        type: string
                        example: "Active"
                      createdAt:
                        type: string
                        format: date-time
                        example: "2024-12-10T00:00:00Z"
        '404':
          description: Acceso al laboratorio no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: "Acceso al laboratorio no encontrado."
        '500':
          description: Error interno en el servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: "Error al obtener el acceso al laboratorio."
  /access-lab/create:
    post:
      tags:
      - Accesos a Laboratorios
      summary: Crear un nuevo acceso a laboratorio
      description: Crea un acceso a un laboratorio para un usuario con la información proporcionada.
      operationId: createAcessLab
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                  example: "Angelo G"
                  description: Nombre completo de la persona que solicita el acceso
                identification_card:
                  type: string
                  example: "0202486650"
                  description: Número de tarjeta de identificación
                type_access:
                  type: string
                  example: "asasas"
                  description: Tipo de acceso al laboratorio
                id_lab_fk:
                  type: string
                  example: "6ee8bb39-62e8-4bde-9005-d80cbd8524e7"
                  description: ID del laboratorio al que se solicita el acceso
              required:
              - full_name
              - identification_card
              - type_access
              - id_lab_fk
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                  example: "Angelo G"
                  description: Nombre completo de la persona que solicita el acceso
                identification_card:
                  type: string
                  example: "0202486650"
                  description: Número de tarjeta de identificación
                type_access:
                  type: string
                  example: "asasas"
                  description: Tipo de acceso al laboratorio
                id_lab_fk:
                  type: string
                  example: "6ee8bb39-62e8-4bde-9005-d80cbd8524e7"
                  description: ID del laboratorio al que se solicita el acceso
      responses:
        '201':
          description: Acceso a laboratorio creado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  code:
                    type: integer
                    example: 201
                  message:
                    type: string
                    example: "Acceso a laboratorio creado exitosamente."
                  data:
                    type: object
                    properties:
                      id_access_lab:
                        type: integer
                        example: 1
                      full_name:
                        type: string
                        example: "Angelo G"
                      identification_card:
                        type: string
                        example: "0202486650"
                      type_access:
                        type: string
                        example: "asasas"
                      id_lab_fk:
                        type: string
                        example: "6ee8bb39-62e8-4bde-9005-d80cbd8524e7"
                      createdAt:
                        type: string
                        format: date-time
                        example: "2024-12-10T00:00:00Z"
                      updatedAt:
                        type: string
                        format: date-time
                        example: "2024-12-10T00:00:00Z"
        '400':
          description: Datos de entrada inválidos o error de validación
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: "Error en los datos proporcionados."
        '404':
          description: Laboratorio no encontrado o no disponible
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: "Laboratorio no encontrado o no disponible."
        '500':
          description: Error interno en el servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: "Error al crear el acceso a laboratorio."
  /access-lab/update/{id}:
    put:
      tags:
      - Accesos a Laboratorios
      summary: Actualizar acceso a laboratorio
      description: Actualiza la información del acceso a un laboratorio para un usuario con los datos proporcionados.
      operationId: updateAcessLab
      parameters:
      - name: id
        in: path
        required: true
        description: El ID del acceso al laboratorio a actualizar
        schema:
          type: string
          example: "6ee8bb39-62e8-4bde-9005-d80cbd8524e7"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                  example: "Angelo G"
                  description: Nombre completo de la persona que solicita el acceso
                identification_card:
                  type: string
                  example: "0202486650"
                  description: Número de tarjeta de identificación
                type_access:
                  type: string
                  example: "asasas"
                  description: Tipo de acceso al laboratorio
                id_lab_fk:
                  type: string
                  example: "6ee8bb39-62e8-4bde-9005-d80cbd8524e7"
                  description: ID del laboratorio al que se solicita el acceso
              required:
              - full_name
              - identification_card
              - type_access
              - id_lab_fk
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                  example: "Angelo G"
                  description: Nombre completo de la persona que solicita el acceso
                identification_card:
                  type: string
                  example: "0202486650"
                  description: Número de tarjeta de identificación
                type_access:
                  type: string
                  example: "asasas"
                  description: Tipo de acceso al laboratorio
                id_lab_fk:
                  type: string
                  example: "6ee8bb39-62e8-4bde-9005-d80cbd8524e7"
                  description: ID del laboratorio al que se solicita el acceso
      responses:
        '200':
          description: Acceso a laboratorio actualizado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "Acceso a laboratorio actualizado exitosamente."
                  data:
                    type: object
                    properties:
                      id_access_lab:
                        type: integer
                        example: 1
                      full_name:
                        type: string
                        example: "Angelo G"
                      identification_card:
                        type: string
                        example: "0202486650"
                      type_access:
                        type: string
                        example: "asasas"
                      id_lab_fk:
                        type: string
                        example: "6ee8bb39-62e8-4bde-9005-d80cbd8524e7"
                      createdAt:
                        type: string
                        format: date-time
                        example: "2024-12-10T00:00:00Z"
                      updatedAt:
                        type: string
                        format: date-time
                        example: "2024-12-10T00:00:00Z"
        '400':
          description: Datos de entrada inválidos o error de validación
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: "Error en los datos proporcionados."
        '404':
          description: Acceso a laboratorio no encontrado o laboratorio no disponible
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: "Acceso a laboratorio no encontrado o laboratorio no disponible."
        '500':
          description: Error interno en el servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: "Error al actualizar el acceso a laboratorio."
...
  /access-lab/report/pdf:
    get:
      tags:
      - Accesos a Laboratorios
      summary: Generar un reporte en PDF de accesos a laboratorios
      description: Genera un reporte en PDF con los accesos a los laboratorios entre dos fechas proporcionadas.
      operationId: generatePdfReport
      parameters:
      - in: query
        name: startDate
        schema:
          type: string
          format: date
          example: "2024-01-01"
        description: Fecha de inicio del rango de fecha para el reporte (en formato YYYY-MM-DD).
      - in: query
        name: endDate
        schema:
          type: string
          format: date
          example: "2024-12-31"
        description: Fecha de fin del rango de fecha para el reporte (en formato YYYY-MM-DD).
      responses:
        '200':
          description: Reporte PDF generado exitosamente
          content:
            application/pdf:
              schema:
                type: string
                format: binary
                example: "PDF generado"
        '400':
          description: Error de validación de parámetros de entrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: "Error en los parámetros proporcionados."
        '404':
          description: No se encontraron registros de accesos para generar el reporte
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: "No se encontraron registros de accesos para generar el reporte."
        '500':
          description: Error interno en el servidor al generar el reporte
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: "Error al generar el reporte PDF de accesos."
  # REACTIVE
...